+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 [![Donate via PayPal](https://img.shields.io/badge/Donate-PayPal-blue.svg?style=flat-square)](https://www.paypal.me/techtoday) 
 
 <p>
+<b>Version 1.0.65</b> September 2022<br/>
+- NEW: new AX Pro and AX Pro Hybrid node. These panels require a totally different authentication and handling, so i made another node just for that. You can receive events and also control your security system<br/>
+- EDIT: changed the node name from "Alarm" and "RAW Alarm" to "Event" and "RAW Event" to avoid confusion with the new AX Pro node.<br/>
+</p>
+<p>
 <b>Version 1.0.64</b> September 2022<br/>
 - Not all Hikvision door intercom does have the proper APIs to connect to. Added warning.<br/>
 - Add: The Picture Node now passes through the input message, in the property msg.previousInputMessage<br/>